Widen lambda input argument since it's wider internally anyway Fix some other small issues a last ai scan discovered

+++ src/Analyse.cpp
@@ -24,7 +24,7 @@ struct AnalyseData {
 
     int deltaFrame;
 
-    int nLambda;
+    int64_t nLambda;
 
     SearchType searchType;
 
@@ -227,7 +227,7 @@ static void VS_CC analyseCreate(const VSMap *in, VSMap *out, [[maybe_unused]] vo
         int pixelMax = (1 << std::min(16, d->vi->format.bitsPerSample)) - 1; // float SAD uses the 16-bit scale
         d->lsad = (int64_t)((double)d->lsad * pixelMax / 255.0 + 0.5);
         d->badSAD = (int64_t)((double)d->badSAD * pixelMax / 255.0 + 0.5);
-        d->nLambda = (int)((double)d->nLambda * pixelMax / 255.0 + 0.5);
+        d->nLambda = (int64_t)((double)d->nLambda * pixelMax / 255.0 + 0.5);
 
         d->lsad = (int64_t)d->lsad * (d->nBlkSizeX * d->nBlkSizeY) / 64;
         d->badSAD = d->badSAD * (d->nBlkSizeX * d->nBlkSizeY) / 64;
